refactor: split 4 files exceeding 800-line limit into submodules
Extract test modules from four files that violated the RUST/file-too-long
limit (800 lines):

- harmonia-db: play_history.rs (1575 lines) → play_history/{mod,tests}.rs
- paroche: opds/catalog.rs (1161 lines) → opds/catalog/{mod,tests}.rs
- harmonia-host: tests/acquisition_integration.rs (1116 lines) →
  extraction of pipeline_tests and recovery_tests into submodule files
- komide: service.rs (817 lines) → service/{mod,tests}.rs

All public APIs unchanged. Zero RUST/file-too-long violations remain.
All tests pass.

Closes #1613
