+++ crates/harmonia-host/tests/acquisition_integration.rs
@@ -6,7 +6,6 @@
 
 use std::pin::Pin;
 use std::sync::Arc;
-use std::time::Duration;
 
 use axum::body::Body;
 use axum::http::{Request, StatusCode};
@@ -18,13 +17,13 @@ use uuid::Uuid;
 
 use ergasia::{DownloadProgress, DownloadState, ErgasiaError, ExtractionResult};
 use exousia::{AuthService, CreateUserRequest, ExousiaServiceImpl, UserRole};
-use harmonia_common::ids::{DownloadId, ReleaseId, WantId};
-use harmonia_common::{HarmoniaEvent, create_event_bus};
+use harmonia_common::ids::DownloadId;
+use harmonia_common::create_event_bus;
 use harmonia_db::DbPools;
 use harmonia_db::migrate::MIGRATOR;
-use horismos::{Config, ExousiaConfig, SyntaxisConfig};
+use horismos::{Config, ExousiaConfig};
 use paroche::state::{AppState, DynSearchService, ServiceFut};
-use syntaxis::{CompletedDownload, ImportService, QueueItem, QueueManager, SyntaxisService};
+use syntaxis::{CompletedDownload, ImportService};
 
 // ── Mock search service ──────────────────────────────────────────────────────
 
@@ -708,409 +707,7 @@ async fn member_on_admin_routes_returns_403() -> Result<(), TestError> {
     Ok(())
 }
 
-// ── Pipeline integration tests (SyntaxisService + MockEngine) ────────────────
-
-fn test_syntaxis_config() -> SyntaxisConfig {
-    SyntaxisConfig {
-        max_concurrent_downloads: 5,
-        max_per_tracker: 3,
-        retry_count: 2,
-        retry_backoff_base_seconds: 0,
-        stalled_download_timeout_hours: 24,
-    }
-}
-
-fn make_queue_item(priority: u8) -> QueueItem {
-    QueueItem {
-        id: Uuid::now_v7(),
-        want_id: WantId::new(),
-        release_id: ReleaseId::new(),
-        download_url: format!("magnet:?xt=urn:btih:{}", Uuid::now_v7()),
-        protocol: syntaxis::DownloadProtocol::Torrent,
-        priority,
-        tracker_id: None,
-        info_hash: None,
-    }
-}
-
-#[tokio::test]
-async fn pipeline_enqueue_dispatches_to_engine() -> Result<(), TestError> {
-    let pool = test_db().await?;
-    let (started_tx, mut started_rx) = mpsc::unbounded_channel();
-    let (imported_tx, _imported_rx) = mpsc::unbounded_channel();
-
-    let engine = Arc::new(MockEngine { started_tx });
-    let import_svc: Arc<dyn ImportService> = Arc::new(MockImportService { imported_tx });
-    let config = test_syntaxis_config();
-
-    let svc = Arc::new(SyntaxisService::new(pool, engine, import_svc, config).await?);
-
-    // Enqueue at priority 4 (interactive bypass) to trigger immediate dispatch
-    let item = make_queue_item(4);
-    let pos = svc.enqueue(item).await?;
-    assert_eq!(pos.position, 0);
-
-    // Wait for the spawned dispatch task to call start_download
-    let dl_id = tokio::time::timeout(Duration::from_secs(5), started_rx.recv())
-        .await?
-        .expect("engine should have received start_download");
-
-    // Verify we got a valid download ID back
-    assert!(!dl_id.to_string().is_empty());
-    Ok(())
-}
-
-#[tokio::test]
-async fn pipeline_completion_triggers_import() -> Result<(), TestError> {
-    let pool = test_db().await?;
-    let (started_tx, mut started_rx) = mpsc::unbounded_channel();
-    let (imported_tx, mut imported_rx) = mpsc::unbounded_channel();
-    let (event_tx, _) = create_event_bus(64);
-
-    let engine = Arc::new(MockEngine { started_tx });
-    let import_svc: Arc<dyn ImportService> = Arc::new(MockImportService { imported_tx });
-    let config = test_syntaxis_config();
-
-    let svc = Arc::new(SyntaxisService::new(pool, engine, import_svc, config).await?);
-    let shutdown = tokio_util::sync::CancellationToken::new();
-    svc.start(event_tx.subscribe(), shutdown.clone());
-
-    // Enqueue at priority 4 to dispatch immediately
-    svc.enqueue(make_queue_item(4)).await?;
-
-    // Wait for engine to receive the download
-    let dl_id = tokio::time::timeout(Duration::from_secs(5), started_rx.recv())
-        .await?
-        .expect("engine should have received start_download");
-
-    // Simulate download completion via event bus
-    event_tx.send(HarmoniaEvent::DownloadCompleted {
-        download_id: dl_id,
-        path: std::path::PathBuf::from("/tmp/test-download"),
-    })?;
-
-    // Wait for import to be triggered
-    let imported_id = tokio::time::timeout(Duration::from_secs(5), imported_rx.recv())
-        .await?
-        .expect("import service should have been called");
-
-    assert_eq!(imported_id.to_string(), dl_id.to_string());
-
-    shutdown.cancel();
-    Ok(())
-}
-
-#[tokio::test]
-async fn pipeline_priority_ordering_in_queue() -> Result<(), TestError> {
-    let pool = test_db().await?;
-    let (started_tx, _started_rx) = mpsc::unbounded_channel();
-    let (imported_tx, _imported_rx) = mpsc::unbounded_channel();
-
-    let engine = Arc::new(MockEngine { started_tx });
-    let import_svc: Arc<dyn ImportService> = Arc::new(MockImportService { imported_tx });
-    // Set max_concurrent to 0 so nothing dispatches (all items stay queued)
-    let config = SyntaxisConfig {
-        max_concurrent_downloads: 0,
-        max_per_tracker: 0,
-        retry_count: 2,
-        retry_backoff_base_seconds: 0,
-        stalled_download_timeout_hours: 24,
-    };
-
-    let svc = Arc::new(SyntaxisService::new(pool, engine, import_svc, config).await?);
-
-    // Enqueue items at different priorities
-    svc.enqueue(make_queue_item(1)).await?;
-    svc.enqueue(make_queue_item(3)).await?;
-    svc.enqueue(make_queue_item(2)).await?;
-
-    let snapshot = svc.get_queue_state().await?;
-    assert_eq!(snapshot.queued_items.len(), 3);
-    // Items ordered by priority: 3, 2, 1 (highest first)
-    assert_eq!(snapshot.queued_items[0].priority, 3);
-    assert_eq!(snapshot.queued_items[1].priority, 2);
-    assert_eq!(snapshot.queued_items[2].priority, 1);
-    Ok(())
-}
-
-#[tokio::test]
-async fn pipeline_fifo_within_same_priority_tier() -> Result<(), TestError> {
-    let pool = test_db().await?;
-    let (started_tx, _started_rx) = mpsc::unbounded_channel();
-    let (imported_tx, _imported_rx) = mpsc::unbounded_channel();
-
-    let engine = Arc::new(MockEngine { started_tx });
-    let import_svc: Arc<dyn ImportService> = Arc::new(MockImportService { imported_tx });
-    let config = SyntaxisConfig {
-        max_concurrent_downloads: 0,
-        max_per_tracker: 0,
-        retry_count: 2,
-        retry_backoff_base_seconds: 0,
-        stalled_download_timeout_hours: 24,
-    };
-
-    let svc = Arc::new(SyntaxisService::new(pool, engine, import_svc, config).await?);
-
-    let item_a = make_queue_item(2);
-    let item_b = make_queue_item(2);
-    let id_a = item_a.id;
-    let id_b = item_b.id;
-
-    svc.enqueue(item_a).await?;
-    svc.enqueue(item_b).await?;
-
-    let snapshot = svc.get_queue_state().await?;
-    assert_eq!(snapshot.queued_items.len(), 2);
-    // FIFO: first enqueued first
-    assert_eq!(snapshot.queued_items[0].id, id_a);
-    assert_eq!(snapshot.queued_items[1].id, id_b);
-    Ok(())
-}
-
-#[tokio::test]
-async fn pipeline_transient_failure_triggers_retry() -> Result<(), TestError> {
-    let pool = test_db().await?;
-    let (started_tx, mut started_rx) = mpsc::unbounded_channel();
-    let (imported_tx, _imported_rx) = mpsc::unbounded_channel();
-    let (event_tx, _) = create_event_bus(64);
-
-    let engine = Arc::new(MockEngine { started_tx });
-    let import_svc: Arc<dyn ImportService> = Arc::new(MockImportService { imported_tx });
-    let config = SyntaxisConfig {
-        max_concurrent_downloads: 5,
-        max_per_tracker: 3,
-        retry_count: 3,
-        retry_backoff_base_seconds: 0,
-        stalled_download_timeout_hours: 24,
-    };
-
-    let svc = Arc::new(SyntaxisService::new(pool.clone(), engine, import_svc, config).await?);
-    let shutdown = tokio_util::sync::CancellationToken::new();
-    svc.start(event_tx.subscribe(), shutdown.clone());
-
-    let item = make_queue_item(4);
-    let queue_id = item.id;
-    svc.enqueue(item).await?;
-
-    // Wait for dispatch
-    let dl_id = tokio::time::timeout(Duration::from_secs(5), started_rx.recv())
-        .await?
-        .expect("engine should start download");
-
-    // Send transient failure (network error, not in permanent patterns)
-    event_tx.send(HarmoniaEvent::DownloadFailed {
-        download_id: dl_id,
-        reason: "connection timeout".to_string(),
-    })?;
-
-    // Wait for retry processing
-    tokio::time::sleep(Duration::from_millis(200)).await;
-
-    // Verify retry_count was incremented in DB and status reset to queued
-    let row: (i64, String) =
-        sqlx::query_as("SELECT retry_count, status FROM download_queue WHERE id = ?")
-            .bind(queue_id.as_bytes().as_slice())
-            .fetch_one(&pool)
-            .await?;
-    assert_eq!(
-        row.0, 1,
-        "retry_count should be 1 after first transient failure"
-    );
-    assert_eq!(
-        row.1, "queued",
-        "status should be reset to queued for retry"
-    );
-
-    shutdown.cancel();
-    Ok(())
-}
-
-#[tokio::test]
-async fn pipeline_permanent_failure_marks_failed() -> Result<(), TestError> {
-    let pool = test_db().await?;
-    let (started_tx, mut started_rx) = mpsc::unbounded_channel();
-    let (imported_tx, _imported_rx) = mpsc::unbounded_channel();
-    let (event_tx, _) = create_event_bus(64);
-
-    let engine = Arc::new(MockEngine { started_tx });
-    let import_svc: Arc<dyn ImportService> = Arc::new(MockImportService { imported_tx });
-    let config = test_syntaxis_config();
-
-    let svc = Arc::new(SyntaxisService::new(pool.clone(), engine, import_svc, config).await?);
-    let shutdown = tokio_util::sync::CancellationToken::new();
-    svc.start(event_tx.subscribe(), shutdown.clone());
-
-    let item = make_queue_item(4);
-    let queue_id = item.id;
-    svc.enqueue(item).await?;
-
-    let dl_id = tokio::time::timeout(Duration::from_secs(5), started_rx.recv())
-        .await?
-        .expect("engine should start download");
-
-    // Send permanent failure (contains "no seeders" which matches permanent pattern)
-    event_tx.send(HarmoniaEvent::DownloadFailed {
-        download_id: dl_id,
-        reason: "no seeders available after 24 hours".to_string(),
-    })?;
-
-    tokio::time::sleep(Duration::from_millis(200)).await;
-
-    let row: (String, Option<String>) =
-        sqlx::query_as("SELECT status, failed_reason FROM download_queue WHERE id = ?")
-            .bind(queue_id.as_bytes().as_slice())
-            .fetch_one(&pool)
-            .await?;
-    assert_eq!(row.0, "failed");
-    assert!(row.1.as_deref().unwrap_or("").contains("no seeders"));
-
-    shutdown.cancel();
-    Ok(())
-}
-
-#[tokio::test]
-async fn pipeline_retry_budget_exhaustion_marks_failed() -> Result<(), TestError> {
-    // NOTE: SyntaxisService has a bug where ActiveEntry.retry_count is always
-    // initialised to 0 regardless of how many retries have occurred in the DB.
-    // This means the in-memory retry_count check (`retry_count >= max_retries`)
-    // only works when max_retries is 0. We set retry_count=0 in config so
-    // the very first transient failure immediately exhausts the budget.
-    let pool = test_db().await?;
-    let (started_tx, mut started_rx) = mpsc::unbounded_channel();
-    let (imported_tx, _imported_rx) = mpsc::unbounded_channel();
-    let (event_tx, _) = create_event_bus(64);
-
-    let engine = Arc::new(MockEngine { started_tx });
-    let import_svc: Arc<dyn ImportService> = Arc::new(MockImportService { imported_tx });
-    let config = SyntaxisConfig {
-        max_concurrent_downloads: 5,
-        max_per_tracker: 3,
-        retry_count: 0,
-        retry_backoff_base_seconds: 0,
-        stalled_download_timeout_hours: 24,
-    };
-
-    let svc = Arc::new(SyntaxisService::new(pool.clone(), engine, import_svc, config).await?);
-    let shutdown = tokio_util::sync::CancellationToken::new();
-    svc.start(event_tx.subscribe(), shutdown.clone());
-
-    let item = make_queue_item(4);
-    let queue_id = item.id;
-    svc.enqueue(item).await?;
-
-    let dl_id = tokio::time::timeout(Duration::from_secs(5), started_rx.recv())
-        .await?
-        .expect("engine should start download");
-
-    // Transient failure with zero retries allowed → immediate budget exhaustion
-    event_tx.send(HarmoniaEvent::DownloadFailed {
-        download_id: dl_id,
-        reason: "connection reset".to_string(),
-    })?;
-
-    tokio::time::sleep(Duration::from_millis(200)).await;
-
-    let row: (String, Option<String>) =
-        sqlx::query_as("SELECT status, failed_reason FROM download_queue WHERE id = ?")
-            .bind(queue_id.as_bytes().as_slice())
-            .fetch_one(&pool)
-            .await?;
-    assert_eq!(row.0, "failed");
-    assert!(
-        row.1
-            .as_deref()
-            .unwrap_or("")
-            .contains("retry budget exhausted")
-    );
-
-    shutdown.cancel();
-    Ok(())
-}
-
-// ── Startup recovery tests ───────────────────────────────────────────────────
-
-#[tokio::test]
-async fn startup_recovery_loads_queued_items_from_db() -> Result<(), TestError> {
-    let pool = test_db().await?;
-
-    // Insert non-terminal rows directly into DB (simulating prior state)
-    let id_queued = Uuid::now_v7();
-    let id_downloading = Uuid::now_v7();
-    let id_completed = Uuid::now_v7();
-    let want_id = Uuid::now_v7().as_bytes().to_vec();
-    let release_id = Uuid::now_v7().as_bytes().to_vec();
-
-    for (id, status) in [
-        (id_queued, "queued"),
-        (id_downloading, "downloading"),
-        (id_completed, "completed"),
-    ] {
-        sqlx::query(
-            "INSERT INTO download_queue \
-             (id, want_id, release_id, download_url, protocol, priority, status, added_at, retry_count) \
-             VALUES (?, ?, ?, 'magnet:test', 'torrent', 2, ?, '2026-01-01T00:00:00Z', 0)",
-        )
-        .bind(id.as_bytes().as_slice())
-        .bind(&want_id)
-        .bind(&release_id)
-        .bind(status)
-        .execute(&pool)
-        .await?;
-    }
-
-    // Boot SyntaxisService — recovery should load non-terminal items
-    let (started_tx, _started_rx) = mpsc::unbounded_channel();
-    let (imported_tx, _imported_rx) = mpsc::unbounded_channel();
-    let engine = Arc::new(MockEngine { started_tx });
-    let import_svc: Arc<dyn ImportService> = Arc::new(MockImportService { imported_tx });
-    let config = SyntaxisConfig {
-        max_concurrent_downloads: 0,
-        max_per_tracker: 0,
-        retry_count: 3,
-        retry_backoff_base_seconds: 30,
-        stalled_download_timeout_hours: 24,
-    };
-
-    let svc = Arc::new(SyntaxisService::new(pool, engine, import_svc, config).await?);
-
-    let snapshot = svc.get_queue_state().await?;
-    // 'queued' and 'downloading' are non-terminal and should be recovered
-    // 'completed' is terminal and should NOT be recovered
-    assert_eq!(
-        snapshot.queued_items.len(),
-        2,
-        "both 'queued' and 'downloading' rows should be recovered into the in-memory queue"
-    );
-    assert_eq!(snapshot.completed_count, 1);
-    Ok(())
-}
-
-#[tokio::test]
-async fn startup_recovery_visible_via_http_snapshot() -> Result<(), TestError> {
-    let (state, auth, pool) = test_state().await?;
-    let token = admin_token(&auth).await?;
-
-    // Insert a queued row directly
-    let id = Uuid::now_v7();
-    let want_id = Uuid::now_v7().as_bytes().to_vec();
-    let release_id = Uuid::now_v7().as_bytes().to_vec();
-    sqlx::query(
-        "INSERT INTO download_queue \
-         (id, want_id, release_id, download_url, protocol, priority, status, added_at, retry_count) \
-         VALUES (?, ?, ?, 'magnet:test', 'torrent', 3, 'queued', '2026-01-01T00:00:00Z', 0)",
-    )
-    .bind(id.as_bytes().as_slice())
-    .bind(&want_id)
-    .bind(&release_id)
-    .execute(&pool)
-    .await?;
-
-    let app = build_app(state);
-
-    let (_, json) = get_queue_snapshot(&app, &token).await?;
-    let queued = json["data"]["queued"].as_array().unwrap();
-    assert_eq!(queued.len(), 1);
-    assert_eq!(queued[0]["priority"], 3);
-    assert_eq!(queued[0]["status"], "queued");
-    Ok(())
-}
+#[path = "acquisition_integration/pipeline_tests.rs"]
+mod pipeline_tests;
+#[path = "acquisition_integration/recovery_tests.rs"]
+mod recovery_tests;
